requirements.txt
pytest
selenium
webdriver-manager
requests
flake8
allure-pytest